EquationSearch: Better search, better expression matching
In this change I introduce a meet-in-the-middle search to EquationSearch
that allows it to search from both hypotheses and goal that makes it more
efficient. I also correct an obvious oversight about how it matches
expression matching to allow it to match more often.

This change adds ~150k transitive implications.