Add first small proofs by MagmaEgg, a Rust egg-based automated prover, proving ~1000 implications, 2.8M -> 1.1M unknown (#216)
This is a simple but apparently at least somewhat effective Rust theorem
prover based on egg e-graph library written for this project.

It proved 5574 of the 24283 implications in the "only_strongest.txt"
file (may be out of date, so some (or most?) could no longer be
relevant). This pull requests only includes 3277 implications with a
Lean proof of up to 100 lines, since the longer ones are very slow to
check in Lean. The "magmaegg-large" branch has the large ones as well.

The code was originally based on the magma_search pull request, but has
been pretty much completely rewritten.

Currently search just uses the egg library in a basic fashion, except
that in case there are extra variables not present in the LHS, it has
code to instantiate them with all subexpressions of the original goal.

Exporting the proofs to Lean has turned out to be harder than finding
the proofs, but i managed to write a solution that seems good (modulo
some issues in egg that require to sometimes turn off explanation
optimization since it sometimes triggers stack overflows and assert
failures) and directly produces proof terms using let, Eq.refl, Eq.symm,
Eq.trans, Magma.op, a congruence lemma for Magma.op and variables and
the hypothesis. I define one letter aliases for them to reduce
verbosity.

Possible future work:
- Figure out which implications are important to prove and try it on
them
- Replace the fork-based code with self-execution so that it works on
Windows and is less of na hack
- Fork egg and fix the buggy optimized explanation generation
- Maybe write Lean code directly instead of writing explanation sexps
and converting to Lean code in a second run
- Fix the generation of extra variable values so it doesn't take too
much time in pathological cases (i.e. goals with 4-6 variables)
- Determine whether it actually has some advantages compared to Vampire
and lean-egg
- Stop the search early if the lhs has been found equal to the rhs
- Support searching for multiple goal equations at once
- Write a custom elaborator for Lean to speed up elaboration
- Support having extra rewrite rules, such as other implications that
have been found implied by the hypothesis, or simple equalities found by
the egraph search itself
- Run it with massive computing resources if deemed useful and someone
offers those