A new tool to find "powerful" theorems (#270)
This is a tool that will find "powerful" theorems. (Theorems that, if
proven, would give many free edges on the graph by implications).

It is slow the first time it runs, because it has to invoke lean to find
all the implications that are proven or conjectured. You can cache the
result of this computation, and after that, it takes ~2 seconds to find
the next best edge to add. Alternatively, you can also ask it for the
top k implications and that takes basically an identical amount of time.
You can also ask for only one equation per equivalence class.

Right now it basically just tells you to prove anything that would imply
Equation 2 because that gives you 1000+ edges for free. But in the
future it may give more useful outputs when there's not so much low
hanging fruit.

It works by finding the sizes of all equivalence classes, and then
trying, in order, the classes from largest to smallest until it finds
one with a missing edge between the two sets. This is fairly fast right
now. It may become slow if we prune all the easy stuff, if so I can
speed it up later.

```
usage: find_powerful_theorems.py [-h] [--topk TOPK] [--one-per-equiv] [--load-cache LOAD_CACHE] [--save-cache SAVE_CACHE]

Find the most useful implications in a matrix.

optional arguments:
  -h, --help            show this help message and exit
  --topk TOPK           Return the k best theorems to prove
  --one-per-equiv       Return only one per equivalence class
  --load-cache LOAD_CACHE
                        Load output from the cache file
  --save-cache SAVE_CACHE
                        Save the output to a cache file
```