Free magma universal properties (#155)
Added univeral properties of `evalInMagma` and `fmapFreeMagma`. (The
universal property of the latter follows directly from that of the
former!)

Closes #133.

I've written my own `isMagmaHom` predicate for the sake of this issue,
but it will need to be replaced with the "official" `MagmaHom`
implementation from #114 once it is complete. So this PR should not be
merged until that issue is resolved, and the necessary tweaks are made.

---------

Co-authored-by: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-authored-by: Shrys <shreyasss94@gmail.com>