Add more lemmas in Confluence and change it slightly
This adds a bunch of lemmas in Confluence and makes two slight changes.

First change is we define NF in terms of rw instead of bu, which seems
more natural. It is equivalent to the other definition, which is retained
as buNF, with a proof that NF <-> buNF.

Second change, bu is changed to now call rw for Leaf as well. This seems more
natural and is also needed to have NF <-> buNF. Note that if rw is a
projection, then nothing changes, since it must be the identity on leaves.