improved equation command (#448)
This makes a few improvements to the docstring and hover behavior of the
`equation` command. Specifically, you can now hover on `equation 37` to
see the type of the equation, and the equation itself is also added to
the hover for `Equation37` (instead of just the type, which is not
particularly useful).