FiniteMagmaExplorer: initial integrated release (#332)
Previously, FiniteMagmaExplorer was developed in an external repository.
This integrates FiniteMagmaExplorer into the main project repostiory,
to maximize compatibility, maintainability, and ensure integrated future
development, as discussed in #332.

There is a built copy included (in dist/) which CI can use to build the
website, and sources in src.

The main script intended for CI is `tools/fme/generate_fme_website.sh`.