All unknowns 1701 -> ... resolved (#557)
Resolves all 13 unknowns with antecedent 1701 negatively. Resolves 30
unknowns altogether.

Uses the modified translation-invariant magma construction, starting
from pretty much the same base model as used for 1661 (see #511).

Refuting 1701 -> 4587 requires a fairly complicated countermodel of this
sort. The other two countermodels are much simpler: 11 of the 13
implications did not even require patching beyond that required for
truncation!

*Note: the dual proofs are not included, so the effective number settled
by the construction is double that number. I'll create an issue for
proving the duals once this is merged.*