Add Functionality to Equation Explorer to Download Raw Implications Table (#628)
Added button as to resolve:
https://github.com/teorth/equational_theories/issues/626

Note that this takes a while to download since its 50MB, as such I
refactored the popup code a little bit to make it more modular so it
could be used for downloads as well. Also, I realize its inefficient to
rematch back to integers but the values are slightly different and
wanted to keep it with what was being asked in 626.

Also provided bugfix for searching by number. Bit of a woopsie on my end
typing out the wrong variable names