`FreeMagmaWithLaws` to `LiftingMagmaFamily` (#635)
The
[two](https://github.com/teorth/equational_theories/blob/main/equational_theories/LiftingMagmaFamilies.lean)
[files](https://github.com/teorth/equational_theories/blob/main/equational_theories/LiftingMagmaFamiliesCounterexamples.lean)
on lifting magma families set up a framework for automatically
generating proofs of non-implications using certain special kinds of
magmas as counter-examples.

The definition of a lifting magma family is tailor-made to this specific
situation and isn't mathematically interesting or natural; this can make
it cumbersome to re-use this framework to generate further proofs of
non-implications.

However, it turns out that any FreeMagmaWithLaws that has DecidableEq
(decidable equality) on its elements automatically gives rise to a
LiftingMagmaFamily. By adding a typeclass instance translating from
FreeMagmaWithLaws with DecidableEq to a LiftingMagmaFamily, it becomes
possible to generate further proofs of non-implications simply by
defining an appropriate FreeMagmaWithLaws, which is a far more natural
definition to deal with mathematically.