temporarily suspending some conjectures to record FME instructions (#716)
I found that if I recorded a finite magma counterexample as a conjecture, then Finite Magma Explorer no longer records the instructions on how to PR the counterexample.  I'm therefore temporarily suspending two conjectures to grab the instructions, put them in the relevant issue, and then restore the conjectures.  (It's a hack.)