Add support for marking implications with the Finite typeclass (#729)
#702 added support for Finite non-implications, this add support for
Finite implications as well. I update the extract_implications script to
always default to 'non-finite' results, and I've verified that the
outputs of `lake exe extract_implications` outputs for every subcommand
match what they were prior to this change, despite marking a number of
Finite results with `@[equational_result]` in this change [modulo the
two non-finite theorems I added the attribute to as well.]

Note that reviewing by commit might be slightly easier, as the first and
last commits are a trivial update and a simple refactor.