Add a "Show Proof" button to the equation explorer.
There is a back-end change to the @equational_result tag, to save
the line number as well as the file name.

This allows the front-end (in a new show_proofs.html file) to link
directly to the location of the proof. It would probably be more
elegant to show this as a pop-up, but I'm not familiar enough with
web development to do this easily.