Experiment: Speed up LawsComplete using rsimp
this copies a prototype for optimizing terms for kernel reduction using
the simplifier from the lean4 repo
(https://github.com/leanprover/lean4/pull/5839) and tries to use it to
prove `testLawsUpto4_computation` without using `native_decide`.

Up to size 3 magmas works in 10s, but size 4 magmas are out of reach so
far, it seems. Maybe with some more optimizations it can be in reach?