Implication statistics (#792)
Add a script that generates scatter plots with implications statistics.

The script currently generates the plot shared here:
https://leanprover.zulipchat.com/#narrow/channel/458659-Equational/topic/Implication.20Statistics/near/480315760

It can, however, easily be modified or extended to show other
implication statistics. There is already some additional data in the
DataFrame used for generating the plot that can be toyed with.