Update InvariantMetatheorem; add Finset (#791)
Should have done with #785, but better late than never. I figured out
how to run the InvariantMetatheorem generator, which added a bunch of
non-implications for the Bol and Moufang loops. (It also dropped some
now unnecessary theorems, which I guess is a fact of life with automated
transitive reduction.) I added the very easy Finset LiftingMagma as
well, which added an extra 190 minimal rules refuting these.

Edit: Actually, those 190 refutations are all of the form "X does not
imply Equation 26302", where 26302 is the natural central groupoid law,
and "X does not imply Equation 345169", where 345169 is the Sheffer
stroke law, for the exact same set of 95 X's. Not sure what's up with
that.

---------

Co-authored-by: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>