Equation Explorer + Graphiti: Finite Graph support
In this change I generate data for the finite graph for both equation
explorer and graphiti, and plumb support for displaying it there. All
previous links and functionality should work identically.

In order to reduce confusion, I add a green stripe to the left-side of
the page when viewing the finite graph.