bug fix: extract_implications --full-entries (#823)
Right now, `extract_implications raw --full-entries` prints entries,
intermingling results for the finite and general graphs. This can lead
to some downstream data issues, such as the two different proofs
[here](https://teorth.github.io/equational_theories/implications/show_proof.html?3308,3456).

In this change, I change it to only print results for either the finite,
or general graph, and a follow-up change will plumb support for the
finite graph to the equation explorer.