Update SmallMagmas.lean
The list of Magmas in SmallMagmas was both incomplete and redundant. (It looked like it was generated with an incorrect notion of isomorphism, which is permuting inputs to the operation but not the outputs.) This is the correct list, and there's a theorem that proves that the list is complete -- powered by several new instances.