Speed up script to find equation id and reorganize it (see #724)
This python code goes from an id to an equation and back, can
generate all equations of a given order.  It is polynomial in the
order in most cases, thanks to a recursive approach based on
knowing how many trees have a given number of nodes on the
left/right, and the same for the rhymes.  For equations that
have the same shape on the left-hand side and right-hand side
the code has to slowly loop through all rhyme schemes to account
for symmetries modulo relabeling, hence takes a time linear in
the equation id in these cases.

It can be loaded as a module, with a better separation between
public and private functions than before.