Correct the max size of (non-1286) non-trivial finite models to 5 (#975)
> In fact our brute force search revealed that in the latter case there
is always a model of order $2 \leq n \leq 4$, with the lone exception of
\eqref{eq1286} (and its dual \eqref{eq2301}), for which the smallest
non-trivial finite model was of order $7$, as presented in
\Cref{1286-ex} below.

4 should be 5 here, we have several magmas with smallest non-trivial
finite model size of 5 (e.g. the outstanding 677).

Actually, even just saying 2-5 kind of understates how dominant 2 is
(counts of {2: 3136, 3: 32, 4: 14, 5: 14, 7: 2} according to
smallest_magma.txt), but with 5 at least it's true. :-)

Co-authored-by: Douglas McNeil <dsm053@gmail.com>