Dashboard: Expand information (#973)
This includes two changes:

1) On every CI run we already generate a lot of graph data that is then
   thrown away, in this change I set it aside, compress it, and upload
   it. Satisfies #631
2) Parse the graph information and output additional information about
   the graph to the dashboard. This includes things like number of
   equivalence classes, number of theorems, and some more. I found a
   need for this while checking some numbers from the paper.

Note to reviewers: This adds a ~120mb blob to the upload step to the
website, I am not sure if this will be an issue, but if it is I am happy
to revert and change things around. I think having the
extract_implications data automatically exported and shared is valuable.