[CONTRIBUTIONS]: Cody Roux (#1037)
Added my own contributions. Validation was the main point, mostly
participating in the proofs of soundness, completeness and compactness,
but also proving that multisets form the free commutative monoid.

Also some Z3 -> Lean countermodel extraction (not sure that was used).

Writing is going to use the blueprint stuff about confluence and
Newmann's lemma.