Correct the numbering of large balanced equations in find_equation_id.py
See [the relevant Zulip thread][1].  The shape comparison used to canonicalize
equations was inconsistent with our preferred order (correctly implemented by
the code producing the list of shapes).  This led to inconsistent mappings
from/to equation ids and incorrect normalization of equation strings.  Now the
output of the following call is correct, instead of swapping lhs and rhs and
giving an unrelated equation id 296850997 corresponding to a different shape.

    $ python find_equation_id.py '((x*x)*x)*(x*x)=(x*((x*x)*x))*x' 296096712
    The equation '((x*x)*x)*(x*x)=(x*((x*x)*x))*x' is Equation 296096712: ((x ◇ x) ◇ x) ◇ (x ◇ x) = (x ◇ ((x ◇ x) ◇ x)) ◇ x
    Equation 296096712: ((x ◇ x) ◇ x) ◇ (x ◇ x) = (x ◇ ((x ◇ x) ◇ x)) ◇ x

[1]: https://leanprover.zulipchat.com/#narrow/channel/458659-Equational/topic/Wrong.20ordering.20of.20equations