Formalize refutations: 1692 -> 47, 1832, 2441, 3050, 3456, 4065 (#1069)
This is based on this paper describing a proof strategy for the above
non-implications:

https://leanprover.zulipchat.com/#narrow/stream/458659-Equational/topic/Proposed.20new.20target.3A.2063.20and.201692.20.28.22Dupont.20and.20Dupond.22.29/near/477252513

There's still quite a bit of room for cleanup (e.g. factoring out common
code from several `match` branches). It's also very likely that there
are simpler ways to prove some of the longer lemmas (e.g.
`partial_function`), as I wasn't working off of the above paper for
those proofs. However, this is now in a state where it's ready for
review - I'm happy to spend more time minimizing this if it would be
useful.

Closes #607

---------

Co-authored-by: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>