Implement progress widget
The dashboard progress visualizer has lost its usefulness: while there were
missing implications, one could track the black pixels, and while it's still
a good way of visualizing the implication/nonimplication graphs, it is too
big to be displayed at full size, and not very useful at small size.

Here I add a small JavaScript widget which replaces the old progress indicator
with an interactive, magnifiable map of the graph.
When clicked, the widget creates a magnifier that the user can move around.
When a particular pixel is selected, the widget opens the Equation Explorer
at the relevant proof.

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>