FRONTEND: Dashboard progress visualization widget (#1181)
**What is this?**

A small JavaScript widget to replace the old progress visualization
image with an interactive, magnifiable version.


![interactive-progress-viz](https://github.com/user-attachments/assets/4f239834-2d92-4c66-a13b-b3b05e0a7497)

When clicked, the widget creates a magnifier that can be moved around. A
status bar explains the implication corresponding to the selected pixel.
When a particular pixel is clicked, the widget gives you a link to open
the Equation Explorer at the corresponding proof.

**Before merging**

Since this changes the website (which I couldn't test myself), if we
decide that we want this merged, it will need to be tested first by one
of the maintainers.

Discussed on
[Zulip](https://leanprover.zulipchat.com/#narrow/channel/458659-Equational/topic/Thoughts.20and.20impressions.20thread/near/512631480).
It also links to an interactive test site where you can try the widget.

Closes #1179

---------

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>