Commentary may (#1248)
Warning: I used Unicode chars such as for the power set in
Equation3385.md and other files. Maybe that's a problem? I also included
links to the equation explorer even for equations beyond order 4. At
present these links do not work, but I figured that they eventually
would and it would be a pain to add them afterwards.

- Commentary for all equations in the longest implication chain(s), some
are interesting on their own rights.
- Commentary on the largest 20 equivalence classes. More than half of
the 4694 laws imply that the operation only depends on one of the two
arguments (or on none).
- Many weaker forms of associativity appearing in the literature.
- Description of free magmas for all laws that imply associativity.
- I discovered many interesting twists of the semi-symmetric quasigroup
law 14 (Putnam law).
- A few loose ends: typos, incorrect mention of Z/3Z modules; law 508,
law 72873.

(The branch name is due to being started in May, nothing deeper.)