ci: restore building the paper PDF (#1250)
Looks accidentally removed in #1249 as part of the rest of the process
being doable via a different action.

I haven't tested this, hopefully it builds on this pull request to see.
I'm slightly worried it won't immediately work, as nothing in this
workflow seems to install latex as a dependency now, but perhaps now
that's hidden behind an earlier action being run, which is why I'd like
to see if it works or not before suggesting merging.