Codyroux/paper canon (#1259)
This is a write up of what is happening in the `Confluence.lean` and
`Confluence_i.lean` developments, for `i = 1...4`, which I attempted to
cast in the classical language of rewrite theory.

This supersedes my previous write up, which was an irrelevant recap of
the classical theory of confluence and convergence in rewrite theory,
which is related but distinct from what is in the Lean.

Closes #926.

Closes #926