Blueprint: change references to laws to E4512 etc. as in the paper
Within chapters we refer to usual LaTeX equations as
'Equation (1.2)' and similar (due to cleveref's \Cref).  To make
equational laws distinct enough I used 'E4512' and the like instead
of 'Equation 4512' as in my previous PR.  Also corrected the web
version (now tested locally).  Hyperlink to the Equation Explorer
instead of to chapter 2 of the blueprint, which was not so useful.
Since plasTeX did not like having a starred version \Equaref*, I
used two separate commands \Equaref, \Equareffull instead.