Forgot to add myself to author block (#1369)
I messed up and only added myself to two of the three places I should
have.

(I noticed this when looking at the author block, it looks like in #1368
@amirlb may have accidentally inserted in the wrong position at the end
instead of alphabetically. I have not fixed that problem because I'm not
sure if that was intentional, but noting that here in case it should be
corrected.)