fix: prevent explorer from displaying any data for tautological equations, and improve error display (#1416)
This PR adds a check to the equation explorer for tautological equations
by checking if the shape and rhyme on the lhs and rhs of the equation
are the same, except for equation1[x=x].

It also removes stack trace from the error pop up in the explorer, but
leaves it in the console instead.


https://github.com/user-attachments/assets/3f165d6e-1275-463d-ae7b-54f82ed2cec3

This resolves #1411

<details>
<summary> Info on use of Number</summary>
In <b>lhsRhyme</b> and <b>rhsRhyme</b> number is used for the slice
method since it does not support <b>bigInt</b>: <a
href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice">https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice</a>
<br>
The code could be re-written to somehow support shapeOrder value which
exceeds JavaScripts <b>Number.MAX_SAFE_INTEGER</b> (2**53), but a string
with the order of the shape on the rhs or lhs exceeding 2*53 will have a
string length of about `2*(2**53) + (2**53 + 1) * 2 + 1` characters,
which is not feasible, hence I just stuck with using <b>Number</b> with
the slice method which is cleaner.
</details>